# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6472ef60c1fa850d0e0744c5
  title: schoolmanagementsystem
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/course/{id}":
    delete:
      x-skaffolder-id: 6472f551c1fa850d0e074594
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6472f17fc1fa850d0e074561
      x-skaffolder-resource: course
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6472f551c1fa850d0e07458b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6472f17fc1fa850d0e074561
      x-skaffolder-resource: course
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
    post:
      x-skaffolder-id: 6472f551c1fa850d0e074588
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6472f17fc1fa850d0e074561
      x-skaffolder-resource: course
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - course
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/course"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/course"
...
  "/exam/{id}":
    delete:
      x-skaffolder-id: 6472f560c1fa850d0e07459d
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6472f55fc1fa850d0e074597
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
    post:
      x-skaffolder-id: 6472f560c1fa850d0e07459a
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exam"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 6472f551c1fa850d0e07458d
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 6472f513c1fa850d0e074574
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 6472f540c1fa850d0e074581
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{is}/validate":
    post:
      x-skaffolder-id: 6472f641c1fa850d0e0745a0
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 6472f1b1c1fa850d0e074567
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: This API is used to validate exam.
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{is}/validate
      x-skaffolder-roles:
      summary: This API is used to validate exam.
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    course:
      x-skaffolder-id: 6472f17fc1fa850d0e074561
      x-skaffolder-id-db: 6472ef60c1fa850d0e0744c6
      x-skaffolder-id-entity: 6472f17ec1fa850d0e07455f
      x-skaffolder-url: /course
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 6472f17ec1fa850d0e074560
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    exam:
      x-skaffolder-id: 6472f1b1c1fa850d0e074567
      x-skaffolder-id-db: 6472ef60c1fa850d0e0744c6
      x-skaffolder-id-entity: 6472f1b1c1fa850d0e074564
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        plaee:
          type: string
          x-skaffolder-id-attr: 6472f1b1c1fa850d0e074566
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 6472f1b1c1fa850d0e074565
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 6472f6e7c1fa850d0e0745a2
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 6472f21dc1fa850d0e07456a
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6472f1b1c1fa850d0e074564
          x-skaffolder-ent2: 6472f17ec1fa850d0e07455f
        _student:
          x-skaffolder-id: 6472f21dc1fa850d0e074569
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6472f1b1c1fa850d0e074564
          x-skaffolder-ent2: 6472f125c1fa850d0e074552
        _teacher:
          x-skaffolder-id: 6472f21dc1fa850d0e07456b
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6472f1b1c1fa850d0e074564
          x-skaffolder-ent2: 6472f161c1fa850d0e074559
      required:
        - _id
...